Rails Broken Asset Checker
This is a small utility that checks for broken assets on your app servers which are sitting behind the Load balance. The use-case is incase that the assets are not compiled on one of our app servers, it will notify on slack with the IP address.

It will use the AWS's API to fetch instances attached to ELBs.
It will then ping them & check if the .CSS & .JS files are returning HTTP 200 Ok